requirements.txt
langgraph
langgraph-prebuilt
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
notebook
tavily-python
wikipedia
trustcall
langgraph-cli[inmem]
numpy
matplotlib
pandas
pydantic
langchain-ollama
scikit-learn
tensorboard
torch-geometric
transformers
openai
e3nn
mendeleev
